Add intermediate snapshot logic for cloning a volume
Recent Embassy policies have impacted our workflow to clone volumes from
existing running volumes as the writing speed (200Mbs) is now hitting timeouts
for our TB scale volumes.
This PR creates an intermediate temporary snapshot that is used as the source
of creation of the new volume leading to response times similar to the ones
that have been used previously.